funckey cleanup: avoid the warning during the delete query when the in is empty

+++ alembic/versions/40e2f31cf146_remove_orphan_func_key.py
@@ -133,6 +133,9 @@ def upgrade():
     )
     func_key_ids = [func_key_result.id for func_key_result in func_key_results]
 
+    if not func_key_ids:
+        return
+
     op.execute(
         func_key_mapping
         .delete()
